refactor(test): split test_task_resume_operations.py into 3 files
Split 525-line test file into 3 focused files (< 300 lines each):
- test_task_resume_reset_scene.py (119 lines, 3 tests)
- test_task_resume_label_state.py (296 lines, 8 tests)
- test_task_resume_core.py (97 lines, 2 tests)

Changes:
- Extract shared fixtures (_make_operations, _mock_label_service) to conftest.py
- All 13 tests preserved and passing

Closes #1341
